move ffi into its own folder
