+++ crates/yomikiri-backend/src/ffi/update_dict.rs
@@ -1,4 +1,5 @@
-use anyhow::{Context, Error};
+use super::backend::RustBackend;
+use super::error::FFIResult;
 use flate2::read::GzDecoder;
 use tempfile::NamedTempFile;
 use yomikiri_dictionary::file::{
@@ -7,88 +8,12 @@ use yomikiri_dictionary::file::{
 };
 use yomikiri_dictionary::metadata::DictMetadata;
 
-use crate::dictionary::Dictionary;
-use crate::error::{YResult, YomikiriError};
-use crate::tokenize::{create_tokenizer, RawTokenizeResult};
-use crate::{utils, SharedBackend};
-use std::fmt;
-use std::fs::{self, File};
+use crate::error::YResult;
+
+use std::fs;
 use std::io::Write;
 use std::path::{Path, PathBuf};
-use std::sync::{Arc, Mutex};
-
-pub type FFIResult<T, E = FFIYomikiriError> = Result<T, E>;
-
-#[derive(Debug, uniffi::Error)]
-pub enum FFIYomikiriError {
-    Error {
-        /// Error message to show to the user
-        message: String,
-        /// Stores list of contexts, starting with `.message` and ending with leaf-most error messsage.
-        details: Vec<String>,
-    },
-}
-
-impl std::error::Error for FFIYomikiriError {}
-
-impl fmt::Display for FFIYomikiriError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Error { message, details } => {
-                write!(f, "{}", self)
-            }
-        }
-    }
-}
-
-#[derive(uniffi::Object)]
-pub struct RustBackend {
-    inner: Mutex<SharedBackend<Vec<u8>, File>>,
-}
-
-impl RustBackend {
-    fn try_from_paths<P1: AsRef<Path>, P2: AsRef<Path>>(
-        index_path: P1,
-        entries_path: P2,
-    ) -> YResult<Arc<RustBackend>> {
-        utils::setup_logger();
-        let tokenizer = create_tokenizer();
-        let dictionary = Dictionary::from_paths(index_path, entries_path)?;
-        let inner = SharedBackend {
-            tokenizer,
-            dictionary,
-        };
-        let inner = Mutex::new(inner);
-        let backend = RustBackend { inner };
-        Ok(Arc::new(backend))
-    }
-}
-
-#[uniffi::export]
-impl RustBackend {
-    #[uniffi::constructor]
-    pub fn new(index_path: String, entries_path: String) -> FFIResult<Arc<RustBackend>> {
-        Self::try_from_paths(&index_path, &entries_path)
-    }
-
-    pub fn tokenize(&self, sentence: String, char_at: u32) -> FFIResult<RawTokenizeResult> {
-        let mut backend = self.inner.lock().unwrap();
-        let char_at = usize::try_from(char_at).map_err(|_| {
-            YomikiriError::ConversionError("Failed to convert char_at to usize".into())
-        })?;
-        let result = backend.tokenize(&sentence, char_at)?;
-        Ok(result)
-    }
-
-    pub fn search(&self, term: String, char_at: u32) -> FFIResult<RawTokenizeResult> {
-        let mut backend = self.inner.lock().unwrap();
-        let char_at = usize::try_from(char_at).map_err(|_| {
-            YomikiriError::ConversionError("Failed to convert char_at to usize".into())
-        })?;
-        let result = backend.search(&term, char_at)?;
-        Ok(result)
-    }
-}
+use std::sync::Arc;
 
 #[derive(uniffi::Object)]
 pub struct DictFilesReplaceJob {
